requirements.txt
streamlit
paddleocr==2.10.0
paddlepaddle -f https://www.paddlepaddle.org.cn/whl/windows/mkl/avx/stable.html
PyMuPDF
pillow
numpy
pandas
openai
tqdm
setuptools
openpyxl